Add Support for Configuration of Database Hostname
I've used this image to migrate a WP installation away from a crummy shared-hosting service. The image was very convenient for getting up and running quickly, however it assumes the use of Compose. When I tried to deploy this to a Cloud66 cluster, I discovered the strict dependence on the  `db` hostname made deployment impossible.

The following change to the entrypoint adds support for a `DB_HOST` environment variable, which defaults to `db` to maintain the existing API, and which is used where the raw `db` hostname was otherwise used.