Add Support for Configuration of Database Hostname (#46)
* Add Support for Configuration of Database Hostname

I've used this image to migrate a WP installation away from a crummy shared-hosting service. The image was very convenient for getting up and running quickly, however it assumes the use of Compose. When I tried to deploy this to a Cloud66 cluster, I discovered the strict dependence on the  `db` hostname made deployment impossible.

The following change to the entrypoint adds support for a `DB_HOST` environment variable, which defaults to `db` to maintain the existing API, and which is used where the raw `db` hostname was otherwise used.

* Updates README

- Adds documentation of the configurability of the database hostname via the
  DB_HOST environment variable.
- Corrects a minor typo.

* Fixes Bug in Entrypoint Script

When a DB_NAME environment variable is not set, the installation of WordPress
will fail silently due to an invalid admin email address. Assigning ADMIN_EMAIL
after null-assigning the local DB_NAME variable ensures a valid email will be
generated by default.